+++ torch_npu/csrc/core/npu/NPUCachingAllocator.cpp
@@ -360,7 +360,7 @@ size_t CachingAllocatorConfig::parseMaxSplitSize(
     size_t i) {
   consumeToken(config, ++i, ':');
   if (++i < config.size()) {
-    size_t val1 = stoi(config[i]);
+    size_t val1 = static_cast<size_t>(stoi(config[i]));
     TORCH_CHECK(
         val1 > kLargeBuffer / (1024 * 1024),
         "CachingAllocator option max_split_size_mb too small, must be > ",
@@ -460,7 +460,7 @@ class DeviceCachingAllocator {
   DeviceCachingAllocator() :
     large_blocks(BlockComparator, false),
     small_blocks(BlockComparator, true) {
-    stats.max_split_size = CachingAllocatorConfig::max_split_size();
+    stats.max_split_size = static_cast<int64_t>(CachingAllocatorConfig::max_split_size());
   }
 
   // All public methods (except the above) acquire the allocator mutex.
@@ -856,7 +856,7 @@ class DeviceCachingAllocator {
 
     const std::array<Block*, 2> merge_candidates = {block->prev, block->next};
     for (Block* merge_candidate : merge_candidates) {
-      const int64_t subsumed_size = try_merge_blocks(block, merge_candidate, pool);
+      const int64_t subsumed_size = static_cast<int64_t>(try_merge_blocks(block, merge_candidate, pool));
       if (subsumed_size > 0) {
         net_change_inactive_split_blocks -= 1;
         net_change_inactive_split_size -= subsumed_size;
@@ -868,7 +868,7 @@ class DeviceCachingAllocator {
 
     if (block->is_split()) {
       net_change_inactive_split_blocks += 1;
-      net_change_inactive_split_size += block->size;
+      net_change_inactive_split_size += static_cast<int64_t>(block->size);
     }
 
     StatTypes stat_types = {false};
@@ -1306,7 +1306,7 @@ class THNCachingAllocator {
   }
 
   void init(int device_count) {
-    int size = device_allocator.size();
+    int size = static_cast<int>(device_allocator.size());
     if (size < device_count) {
       device_allocator.resize(device_count);
       for (const auto i : c10::irange(size, device_count)) {
@@ -1353,13 +1353,13 @@ class THNCachingAllocator {
   }
 
   void emptyCache(bool check_error) {
-    int count = device_allocator.size();
+    int count = static_cast<int>(device_allocator.size());
     for (int i = 0; i < count; i++)
       device_allocator[i]->emptyCache(check_error);
   }
 
   void THNSetShutdownStats() {
-    int count = device_allocator.size();
+    int count = static_cast<int>(device_allocator.size());
     for (int i = 0; i < count; i++)
       device_allocator[i]->devSetShutdownStats();
   }
@@ -1407,7 +1407,7 @@ class THNCachingAllocator {
 
   std::vector<SegmentInfo> snapshot() {
     std::vector<SegmentInfo> result;
-    int count = device_allocator.size();
+    int count = static_cast<int>(device_allocator.size());
     for (int i = 0; i < count; i++) {
       auto snap = device_allocator[i]->snapshot();
       result.insert(result.end(), snap.begin(), snap.end());
@@ -1539,7 +1539,7 @@ void* MallocBlock(size_t size, void *stream, int device) {
   if (device == -1) {
     NPU_CHECK_ERROR(aclrtGetDevice(&device));
   }
-  if ((device < 0) || (device > caching_allocator.device_allocator.size())) {
+  if ((device < 0) || (device > static_cast<int>(caching_allocator.device_allocator.size()))) {
     return nullptr;
   }
   AT_ASSERT(caching_allocator.device_allocator[device]);
